Refactor bond valuation to use double precision
Replaced `decimal` with `double` for numerical properties and calculations across the codebase, including `PresentValue`, `YearsToMaturity`, `FaceValue`, and `DiscountFactor`. Updated methods, tests, and assertions to reflect this change.

Removed the custom `BondValuationException` and replaced it with generic `Exception` handling. Simplified `DeskNotes` handling and improved `Rate.Parse` logic to use `double`.

Refactored and cleaned up test cases, removing redundant methods and improving consistency. Adjusted `CouponBond` and `ZeroCouponBond` logic to use `double` for calculations and removed unnecessary rounding.

Performed general cleanup, including removing unused namespaces, simplifying error messages, and improving code readability.

+++ Domain/BondCreationData.cs
@@ -1,17 +1,18 @@
 namespace Domain
 {
+    //This class matches the structure of the CSV input for bond creation
     public record BondCreationData
     {
         // Unconventional naming to match CSV headers for BondID
         public required string BondID { get; init; }
         public required string Issuer { get; init; }
         public required string Rate { get; init; }
-        public required decimal FaceValue { get; init; }
+        public required double FaceValue { get; init; }
         public required string PaymentFrequency { get; init; }
         public required string Rating { get; init; }
         public required string Type { get; init; }
-        public required decimal YearsToMaturity { get; init; }
-        public required decimal DiscountFactor { get; init; }
+        public required double YearsToMaturity { get; init; }
+        public required double DiscountFactor { get; init; }
         public string? DeskNotes { get; init; }
     }
 }
